Fixed scroll offset adjustment due to inputBarHeight change being done twice sometimes.
When increasing the height constraint of a scrollview and calling layout on it, _adjustContentOffsetIfNecessary is called which adjusts the content offset so that there's no blank space at the bottom. When this happens, the content offset adjustment would happen twice, once by _adjustContentOffsetIfNecessary and once by the code in the block.

Have the adjustment be based on the contentOffset before any layout so that it remains accurate.